requirements.txt
Django==3.2.14
uWSGI==2.0.24
psycopg2-binary
django-crispy-forms==1.14.0
Pillow
gunicorn
celery==5.2.7
redis
flower
djangorestframework
eolymp==0.8.6
protobuf
requests
markdown
django-markdownify
django-celery-beat
django-cors-headers==3.13.0
django-utils-six
django-recaptcha==3.0.0
djangorestframework_simplejwt
drf-recaptcha
drf-yasg
pygments
python-markdown-math
tenacity
httpx
django-storages[s3]
boto3